+++ apps/web/src/app/(main)/(docs)/components/kbd/variants/usage.tsx
@@ -1,0 +1,5 @@
+import { Kbd } from "@jamsr-ui/react";
+
+export const KbdUsage = () => {
+  return <Kbd keys={["command"]}>K</Kbd>;
+};