+++ packages/components/kbd/src/index.ts
@@ -1,0 +1,3 @@
+"use client";
+
+export { Kbd, type KbdProps } from "./kdb";