+++ packages/components/kbd/src/types.d.ts
@@ -1,0 +1,7 @@
+import { type KbdProps } from ".";
+
+declare module "@jamsr-ui/styles" {
+  export interface UIStylesType {
+    kbd?: Partial<KbdProps>;
+  }
+}