+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/input
 
+## 0.0.83
+
+### Patch Changes
+
+- fixes
+
 ## 0.0.82
 
 ### Patch Changes